cleanup

+++ main.py
@@ -1,5 +1,7 @@
 # https://www.notion.so/List-Of-Chains-2497cbaae9c6447b8bdb049e7acd806b
+import os
 from pick import pick
+from CHAINS import VALIDATING_CHAINS
 
 '''
 Reece Williams | Notional - Started Oct 28th 2022
@@ -19,168 +21,10 @@ Goal:
 VERSION BUMPING:
 - Get current version from go.mod (ex: ibc-go v3.3.0)
 - Bump to v3.3.1
+
+TODO:
+- Use mp to run all of these in parallel
 '''
-
-# chains real repos (put notional repos here in the future so we have write access)
-
-# ToDO: if we do not have a repo, we should auto fork it for notional?
-VALIDATING_CHAINS = {
-    # "Osmosis": { # currently protected main
-    #     "branch": "main",
-    #     "root": "git@github.com:osmosis-labs/osmosis.git",
-    #     "notional": "git@github.com:notional-labs/osmosis.git"
-    # },
-    # "Gaia": {
-    #     "branch": "main",
-    #     "root": "git@github.com:cosmos/gaia.git",
-    #     "notional": "git@github.com:notional-labs/gaia.git"
-    # },    
-    # "Juno": { 
-    #     "branch": "main",
-    #     "root": "git@github.com:CosmosContracts/juno.git",
-    #     "notional": "git@github.com:notional-labs/juno.git"
-    # },
-    # "Sifchain": { 
-    #     "branch": "master",
-    #     "root": "git@github.com:Sifchain/sifnode.git",
-    #     "notional": "git@github.com:notional-labs/sifnode.git"
-    # },    
-    # "Omniflix": {
-    #     "branch": "main",
-    #     "root": "git@github.com:OmniFlix/omniflixhub.git",
-    #     "notional": "git@github.com:notional-labs/omniflixhub.git"
-    # },
-    # "Secret": {
-    #     "branch": "master",
-    #     "root": "git@github.com:scrtlabs/SecretNetwork.git",
-    #     "notional": "git@github.com:notional-labs/SecretNetwork.git"
-    # },
-    "Starname-IOV": {
-        "branch": "master",
-        "root": "git@github.com:iov-one/starnamed.git",
-        "notional": "git@github.com:notional-labs/starnamed.git"
-    },
-    "Regen": {
-        "branch": "master",
-        "root": "git@github.com:regen-network/regen-ledger.git",
-        "notional": "git@github.com:notional-labs/regen-ledger.git"
-    },
-    "Akash": {
-        "branch": "master",
-        "root": "git@github.com:ovrclk/akash.git",
-        "notional": "git@github.com:notional-labs/akash.git"
-    },
-    "Sentinel": {
-        "branch": "master",
-        "root": "git@github.com:sentinel-official/sentinel.git",
-        "notional": "git@github.com:notional-labs/sentinel.git"
-    },
-    "Emoney": {
-        "branch": "develop",
-        "root": "git@github.com:e-money/em-ledger.git",
-        "notional": "git@github.com:notional-labs/em-ledger.git"
-    },
-    "Emoney": {
-        "branch": "main",
-        "root": "git@github.com:ixofoundation/ixo-blockchain.git",
-        "notional": "git@github.com:notional-labs/ixo-blockchain.git"
-    },
-    "LikeCoin": {
-        "branch": "master",
-        "root": "git@github.com:likecoin/likecoin-chain.git",
-        "notional": "git@github.com:notional-labs/likecoin-chain.git"
-    },
-    # "Cyber": {
-    #     "branch": "main",
-    #     "root": "git@github.com:cybercongress/go-cyber.git",
-    #     "notional": "git@github.com:notional-labs/go-cyber.git"
-    # },
-    "Cheqd": {
-        "branch": "main",
-        "root": "git@github.com:cheqd/cheqd-node.git",
-        "notional": "git@github.com:notional-labs/cheqd-node.git"
-    },
-    "Stargaze": {
-        "branch": "main",
-        "root": "git@github.com:public-awesome/stargaze.git",
-        "notional": "git@github.com:notional-labs/stargaze.git"
-    },
-    "BandChain": {
-        "branch": "master",
-        "root": "git@github.com:bandprotocol/chain.git",
-        "notional": "git@github.com:notional-labs/chain.git"
-    },
-    "Chihuahua": {
-        "branch": "main",
-        "root": "git@github.com:ChihuahuaChain/chihuahua.git",
-        "notional": "git@github.com:notional-labs/chihuahua.git"
-    },
-    "Kava": {
-        "branch": "master",
-        "root": "git@github.com:Kava-Labs/kava.git",
-        "notional": "git@github.com:notional-labs/kava.git"
-    },
-    "BitCanna": {
-        "branch": "main",
-        "root": "git@github.com:BitCannaGlobal/bcna.git",
-        "notional": "git@github.com:notional-labs/bcna.git"
-    },
-    "Evmos": {
-        "branch": "main",
-        "root": "git@github.com:evmos/evmos.git",
-        "notional": "git@github.com:notional-labs/evmos.git"
-    },
-    "Provenance": {
-        "branch": "main",
-        "root": "git@github.com:provenance-io/provenance.git",
-        "notional": "git@github.com:notional-labs/provenance.git"
-    },
-    "Vidulum": {
-        "branch": "main",
-        "root": "git@github.com:vidulum/mainnet.git",
-        "notional": "git@github.com:notional-labs/vidulum.git"
-    },
-    "Bitsong": {
-        "branch": "main",
-        "root": "git@github.com:bitsongofficial/go-bitsong.git",
-        "notional": "git@github.com:notional-labs/go-bitsong.git"
-    },
-    "FetchAI": {
-        "branch": "master",
-        "root": "git@github.com:fetchai/fetchd.git",
-        "notional": "git@github.com:notional-labs/fetchd.git"
-    },
-    "Uume": {
-        "branch": "main",
-        "root": "git@github.com:umee-network/umee.git",
-        "notional": "git@github.com:notional-labs/umee.git"
-    },
-    "Passage3D": {
-        "branch": "main",
-        "root": "git@github.com:envadiv/Passage3D.git",
-        "notional": "git@github.com:notional-labs/Passage3D.git"
-    },
-    "Stride": {
-        "branch": "main",
-        "root": "git@github.com:Stride-Labs/stride.git",
-        "notional": "git@github.com:notional-labs/stride.git"
-    },
-    "QuickSilver": {
-        "branch": "main",
-        "root": "git@github.com:ingenuity-build/quicksilver.git",
-        "notional": "git@github.com:notional-labs/quicksilver.git"
-    },
-    "Jackal": {
-        "branch": "master",
-        "root": "git@github.com:JackalLabs/canine-chain.git",
-        "notional": "git@github.com:notional-labs/canine-chain.git"
-    },
-    # # "Injective": "",    
-    # # "Ki": "", # ??  
-    # # "Konstellation": "",
-}
-
-import os
 
 # current dir of the file
 current_dir = os.path.dirname(os.path.realpath(__file__))
@@ -192,7 +36,7 @@ def get_chains():
         yield chain, locations
 
 
-def pull_latest(folder_name):
+def pull_latest(folder_name, repo_sync=False):
     # git clone if it is not there already, if it is, cd into dir and git pull    
     os.chdir(current_dir)
 
@@ -212,7 +56,8 @@ def pull_latest(folder_name):
 
     # print(f"Syncing fork to latest of parent main/master branch (overwriting any main/master changes)")    
     os.system(f"git remote add upstream {repo_link}") # if not already
-    os.system(f"gh repo sync {repo_link.replace('git@github.com:', '').replace('.git', '')} --force")
+    if repo_sync:
+        os.system(f"gh repo sync {repo_link.replace('git@github.com:', '').replace('.git', '')} --force")
 
     # pull after we sync to the new files in the notional repo which we will build off of
     os.system("git pull")
@@ -283,43 +128,62 @@ def workflows(folder_name):
 
 
 
-def go_mod_update(folder_name, values: list[str] = []):
+def go_mod_update(folder_name, values: list[str] = [], simulate: bool = False):
     os.chdir(os.path.join(current_dir, folder_name))
     
+    if "go.mod" not in os.listdir():
+        print(f"{folder_name} does not have a go.mod file, skipping...")
+        return
+
     # read go.mod data
     with open("go.mod", "r") as f:
         data = f.read()
         
     for replacements in values:
-        # print(f"Replacing {replacements[0]} with {replacements[1]}")
-        data = data.replace(replacements[0], replacements[1])
-    # print(data)
+        old = replacements[0]
+        new = replacements[1]        
 
-    with open("go.mod", "w") as f:
-        print(f"Writing go.mod for {folder_name}")
-        f.write(data)
+        if old in data:
+            if simulate:            
+                print(f"{folder_name:<15} Would replace {old} w/ {new}")    
+            else:                
+                print(f"Replacing {old} with {new}")
+                data = data.replace(old, new)
 
-    os.system("go mod tidy")
+    if simulate == False:
+        with open("go.mod", "w") as f:
+            print(f"Writing go.mod for {folder_name}")
+            f.write(data)
+
+        os.system("go mod tidy")
     os.chdir(current_dir)
 
 
 
 PATCH_BRANCH_NAME = "reece-ibc331" # this is made on notionals fork
 def oct_28_2022_patches():
+    
+    select_chains = []
     for chain, loc in get_chains():
-        print(f"Chain: {chain}")
-        pull_latest(chain)
+        if len(select_chains) > 0 and chain not in select_chains:
+            continue
+
+        # print(f"Chain: {chain}")
+        # pull_latest(chain, repo_sync=False)
         # dependabot(chain)
         # workflows(chain)
 
-        # go_mod_update(chain, [
-        #     # allow regex matching here? so we could do v3.*.* -> v3.3.1
-        #     ["github.com/cosmos/ibc-go/v3 v3.3.0", "github.com/cosmos/ibc-go/v3 v3.3.1"]
-        # ])
+        go_mod_update(chain, [
+            # allow regex matching here? so we could do v3.*.* -> v3.3.1
+            ["github.com/cosmos/ibc-go/v3 v3.3.0", "github.com/cosmos/ibc-go/v3 v3.3.1"]
+        ], simulate=True)
 
         # lint(chain)
 
 
 
-
-oct_28_2022_patches()
+try:
+    oct_28_2022_patches()
+except Exception as e:
+    print(e)
+    print("Something went wrong, please check the logs above")
